updated some geometries based on first test pcb...
... and added a bunch more footprints showing
locations of target vias, pads, etc.

This should now cover pretty much all of the things,
except for the disk drive which I have no intention
of covering.  Unless maybe it's a standard-ish interface
and can interface to non-Wii disk drives, since that
would make for a pretty badass Wii laptop build.
But that's pretty unlikely.